Reorganize project files and folders
