+++ swiss_student_location/index.html
@@ -8,6 +8,6 @@
   </head>
   <body class="font-inter antialiased text-gray-600 tracking-tight">
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/Main.tsx"></script>
   </body>
 </html>
