+++ swiss_student_location/src/Main.tsx
@@ -7,12 +7,12 @@ import 'primeicons/primeicons.css';
 import 'primeflex/primeflex.css';
 import '@assets/style/index.css'
 import {BrowserRouter, Route, Routes} from "react-router-dom";
-import Home from "@pages/home";
+import Home from "@pages/Home";
 import SignUp from "@pages/users/signUp";
 import SignIn from "@pages/users/signIn";
-import NotFound404 from "@pages/notFound404";
-import Layout from "@pages/layout";
-import Detail from '@pages/detail';
+import NotFound404 from "@pages/NotFound404";
+import Layout from "@pages/Layout";
+import Details from '@pages/details/Details';
 
 export default function App() {
     return (
@@ -22,7 +22,7 @@ export default function App() {
                     <Route path='/' element={<Home/>}/>
                     <Route path='/signin' element={<SignIn/>}/>
                     <Route path='/signup' element={<SignUp/>}/>
-                    <Route path='apartment/:id' element={<Detail/>}/>
+                    <Route path='apartment/:id' element={<Details/>}/>
                     <Route path='*' element={<NotFound404/>}/>
                 </Route>
             </Routes>
