+++ swiss_student_location/src/components/firebase/SignOut.tsx
@@ -1,6 +1,10 @@
 import {auth} from "@config/firebase";
 import { signOut } from "firebase/auth"
 
+/**
+ * Sign out button
+ * @constructor
+ */
 export const SignOut = () => {
     const logOut = async () => {
         await signOut(auth);
