+++ swiss_student_location/src/components/search/RoomSize.tsx
