+++ swiss_student_location/src/models/Canton.ts
