+++ swiss_student_location/src/models/Institution.ts
