+++ swiss_student_location/src/models/RealState.ts
