+++ swiss_student_location/src/models/RealStateType.ts
