+++ swiss_student_location/src/pages/Layout.tsx
@@ -1,5 +1,5 @@
 import {Outlet} from "react-router-dom";
-import Navbar from "@components/navbar";
+import Navbar from "@components/Navbar";
 
 
 export default function Layout() {
